Chore: Update cos tests
