+++ tests/integration/expressjs/conftest.py
@@ -176,59 +176,3 @@ def expressjs_app_fixture(
 @pytest.fixture(autouse=True)
 def cwd():
     return os.chdir(PROJECT_ROOT / "examples/expressjs")
-
-
-@pytest.fixture(scope="module", name="prometheus_app")
-@pytest.mark.skip_juju_version("3.4")
-def deploy_prometheus_fixture(
-    juju: jubilant.Juju,
-    prometheus_app_name: str,
-):
-    """Deploy prometheus."""
-    if not juju.status().apps.get(prometheus_app_name):
-        juju.deploy(
-            prometheus_app_name,
-            channel="1.0/stable",
-            revision=129,
-            base="ubuntu@20.04",
-            trust=True,
-        )
-    juju.wait(
-        lambda status: status.apps[prometheus_app_name].is_active,
-        error=jubilant.any_blocked,
-    )
-
-
-@pytest.fixture(scope="module", name="loki_app")
-@pytest.mark.skip_juju_version("3.4")
-def deploy_loki_fixture(
-    juju: jubilant.Juju,
-    loki_app_name: str,
-):
-    """Deploy loki."""
-    if not juju.status().apps.get(loki_app_name):
-        juju.deploy(loki_app_name, channel="latest/stable", trust=True)
-    juju.wait(
-        lambda status: status.apps[loki_app_name].is_active,
-        error=jubilant.any_blocked,
-    )
-
-
-@pytest.fixture(scope="module", name="cos_apps")
-@pytest.mark.skip_juju_version("3.4")
-def deploy_cos_fixture(
-    juju: jubilant.Juju,
-    loki_app,  # pylint: disable=unused-argument
-    prometheus_app,  # pylint: disable=unused-argument
-    grafana_app_name: str,
-):
-    """Deploy the cos applications."""
-    if not juju.status().apps.get(grafana_app_name):
-        juju.deploy(
-            grafana_app_name,
-            channel="1.0/stable",
-            revision=82,
-            base="ubuntu@20.04",
-            trust=True,
-        )
-    juju.wait(jubilant.all_active)
